code/backend/requirements.txt
tensorflow
numpy
opencv-python-headless
requests
beautifulsoup4
pillow
torch
torchtext
nltk
